fixup! TF-4726 Add web OPFS drive file stager and strategy selection
review-comment: 3763766518

https://github.com/linagora/tmail-flutter/pull/4746#discussion_r3763766518
