+++ workplace/lib/data/datasource/drive_transfer/opfs_drive_file_stager.dart
@@ -3,6 +3,7 @@ import 'dart:math' show Random;
 import 'package:core/utils/app_logger.dart';
 import 'package:core/utils/build_utils.dart';
 import 'package:dio/dio.dart';
+import 'package:workplace/data/datasource/drive_transfer/drive_download_source.dart';
 import 'package:workplace/data/datasource/drive_transfer/drive_file_stager.dart';
 import 'package:workplace/data/datasource/drive_transfer/opfs_fetch_download.dart';
 import 'package:workplace/data/datasource/drive_transfer/opfs_file_handle.dart';
